requirements.txt
pandas
sky-api-python-client-main.zip
google-cloud-bigquery